Latitude and longitude from long to double

+++ app/src/main/java/gr/kalymnos/sk3m3l10/prognosis/common/weather/LocationWeather.java
@@ -9,9 +9,9 @@ import gr.kalymnos.sk3m3l10.prognosis.common.weather_units.WeatherUnits;
 
 public class LocationWeather extends Weather {
 
-    private long lat,lon;
+    private double lat,lon;
 
-    public LocationWeather(long lat,long lon,long timeMilli, String mainWeather, String description,
+    public LocationWeather(double lat,double lon,long timeMilli, String mainWeather, String description,
                            int tempHigh, int tempLow, int humidity, int pressure, double wind,
                            WeatherUnits weatherUnit) {
         super(timeMilli, mainWeather, description, tempHigh, tempLow, humidity, pressure, wind, weatherUnit);
